calc rank use cache
