+++ app/src/main/kotlin/com/example/blogsearcher/app/rank/QueryKeywordRank.kt
@@ -1,0 +1,7 @@
+package com.example.blogsearcher.domain.rank.query
+
+import java.time.LocalDateTime
+
+interface QueryBlogSearchKeywordRank {
+    fun getRank(from: LocalDateTime, to: LocalDateTime): Map<String, Int>
+}
