--- domain/src/main/kotlin/com/example/blogsearcher/domain/rank/BlogSearchKeywordRank.kt
@@ -1,9 +1,0 @@
-package com.example.blogsearcher.domain.rank
-
-import com.example.blogsearcher.domain.search.vo.Keyword
-
-class BlogSearchKeywordRank(
-    val rank: Int,
-    val keyword: Keyword,
-    val searchCount: Long
-)
