--- domain/src/main/kotlin/com/example/blogsearcher/domain/rank/query/QueryBlogSearchKeywordRank.kt
@@ -1,8 +1,0 @@
-package com.example.blogsearcher.domain.rank.query
-
-import com.example.blogsearcher.domain.rank.BlogSearchKeywordRank
-import java.time.LocalDateTime
-
-interface QueryBlogSearchKeywordRank {
-    fun getRank(from: LocalDateTime, to: LocalDateTime): List<BlogSearchKeywordRank>
-}
