+++ infra/src/main/kotlin/com/example/blogsearcher/infra/persistence/repository/QueryKeywordRankHandler.kt
@@ -1,6 +1,5 @@
 package com.example.blogsearcher.infra.persistence.repository
 
-import com.example.blogsearcher.domain.rank.BlogSearchKeywordRank
 import com.example.blogsearcher.domain.rank.query.QueryBlogSearchKeywordRank
 import com.example.blogsearcher.domain.search.vo.Keyword
 import com.example.blogsearcher.infra.persistence.entity.QBlogSearchRecordEntity.blogSearchRecordEntity
@@ -11,14 +10,14 @@ import java.time.LocalDateTime
 
 class BlogSearchRecordRankDto @QueryProjection constructor(
     val keyword: Keyword,
-    val count: Long
+    val count: Int
 )
 
 @Repository
 class QueryBlogSearchKeywordRankHandler(
     private val factory: JPAQueryFactory
 ) : QueryBlogSearchKeywordRank {
-    override fun getRank(from: LocalDateTime, to: LocalDateTime): List<BlogSearchKeywordRank> {
+    override fun getRank(from: LocalDateTime, to: LocalDateTime): Map<String, Int> {
         val result = this.factory.select(QBlogSearchRecordRankDto(blogSearchRecordEntity.keyword, blogSearchRecordEntity.count()))
             .from(blogSearchRecordEntity)
             .where(blogSearchRecordEntity.searchAt.between(from, to))
@@ -26,10 +25,8 @@ class QueryBlogSearchKeywordRankHandler(
             .limit(10)
             .fetch()
 
-        result.sortByDescending { it.count }
-
-        return result.mapIndexed { index, dto ->
-            BlogSearchKeywordRank(index + 1, dto.keyword, dto.count)
-        }
+        return result
+            .sortedByDescending { it.count }
+            .associate { it.keyword.value to it.count }
     }
 }
